Container and hotfixes

See merge request click-n-wash/click-n-wash!72

+++ backend/src/seed.ts
@@ -17,6 +17,15 @@ import { randomDate, randomNumber } from './utils/utils';
  * Seed the database with random data
  */
 const seed = async () => {
+  // Check if DOCKER_SEED is set to true and DB is already seeded
+  if (
+    process.env.DOCKER_SEED === 'true' &&
+    (await getDb().userRepository.findOne({ where: { email: 'seed@gmail.com' } }))
+  ) {
+    console.log('Database already seeded');
+    return;
+  }
+
   // Delete all data from the database
   console.log('Deleting all data from the database...');
   await getDb().dropDatabase();
@@ -43,7 +52,7 @@ const seed = async () => {
     const user = getDb().userRepository.create({
       email: fakerDE.internet.email(),
       name: fakerDE.person.fullName(),
-      balance: 500,
+      balance: 1500,
     });
     userMap.set(user.email, user);
 
@@ -69,7 +78,7 @@ const seed = async () => {
         id: firebaseUser.user.uid,
         email: firebaseUser.user.email!,
         name: 'Seed User',
-        balance: 500,
+        balance: 1500,
       });
       userMap.set(user.email, user);
     }
